test: Use runner in integration_test
